+++ src/components/Organisms/ImpactSlider/ImpactSlider.style.js
@@ -1,10 +1,12 @@
 import styled from 'styled-components';
 import RichText from '../../Atoms/RichText/RichText';
+import Text from '../../Atoms/Text/Text';
+import Button from '../../Atoms/Button/Button';
 
 const OuterWrapper = styled.div`
   position: relative;
   margin: 0 auto;
-  max-width: 1200px;
+  max-width: 100%;
   background-color: ${props => props.theme.color('grey_extra_light')};
   padding: 10px;
 
@@ -13,21 +15,26 @@ const OuterWrapper = styled.div`
   }
 `;
 
+const InnerWrapper = styled.div`
+  width: 100%;
+  max-width: 780px;
+  margin: 0 auto;
+`;
+
 const Copy = styled(RichText)`
   position: relative;
 `;
 
-const CurrentAmountCopy = styled.p`
+const CurrentAmountCopy = styled(Text)`
   font-weight: bold;
-  margin-top: 20px;
+  margin-top: 25px;
 `;
 
-const SubmitButton = styled.button`
+const SubmitButton = styled(Button)`
   margin-top: 20px;
   width: 100%;
-  background-color: red;
 `;
 
 export {
-  OuterWrapper, Copy, CurrentAmountCopy, SubmitButton
+  OuterWrapper, Copy, CurrentAmountCopy, SubmitButton, InnerWrapper
 };
